+++ CHANGELOG.md
@@ -5,13 +5,13 @@
 - defined error codes for what linter will cover
 - added import rules module and parse utils and error codes (#9)
 - added rules checker class to process the results of parsing through file (#8)
-- add custom import rules vistor class for traversing imports (#6)
+- add custom import rules visitor class for traversing imports (#6)
 
 ## v0.1.0 (2023-04-26)
 
 ### Feat
 
-- added parsedimport namedtuple for holding parsed imports from ast
+- added parsed import namedtuple for holding parsed imports from ast
 
 ## v0.0.1 (2023-04-24)